requirements.txt
flask
yt-dlp
colorama
google-generativeai
pyinstaller
